LD54: fix narrow conversion warnings

+++ headers/game/house_generator.h
@@ -30,7 +30,7 @@ namespace game
         int yOff = screenH/2 - houseH/2;
         int wallThickness = 4;
 
-        Rectangle container = { houseW, houseH };
+        Rectangle container = { (float)houseW, (float)houseH };
 
         RoomHandler* roomHandler = nullptr;
     };
